feat: Add click functionality to navbar buttons
This commit adds click functionality to the navbar buttons in the App component. When the "Projects" button is clicked, the user is redirected to the "https://alexkorol.github.io/projects" URL. Similarly, when the "AI Art" button is clicked, the user is redirected to the "https://alexkorol.github.io/aiart" URL. For all other buttons, the setActiveSection function is called to update the active section based on the clicked button.